moved page editor
