+++ client/src/components/FlipBook.tsx
@@ -1,19 +1,10 @@
-import React, { useState, useContext, useRef, createContext, useEffect } from "react";
+import React, { useState, useRef } from "react";
 import HTMLFlipBook from "react-pageflip";
-import STT from './STT';
 import '../styles.css';
-import { FaSave, FaTimesCircle, FaPlusCircle, FaVolumeUp, FaEraser } from 'react-icons/fa';
+import { FaPlusCircle, FaVolumeUp } from 'react-icons/fa';
 import TooltipIcon from './TooltipIcons';
-import { TTSToggleContext } from './Stories';
-// import Switch from "react-switch";
-// import axios from "axios";
-import { GrammarlyEditorPlugin } from "@grammarly/editor-sdk-react";
 import styled from 'styled-components';
-import { io, Socket } from 'socket.io-client';
-
-
-export const socket = io('/');
-export const SocketContext = createContext<Socket | null>(null)
+import PageEditor from "./PageEditor";
 
 const TitlePage: any = styled.div`
   data-density: hard;
@@ -58,128 +49,6 @@ interface Story {
   coverImage: any | null;
   numberOfPages: number | null;
 }
-
-interface PageEditorProps {
-  page: Page;
-  onSave: (content: string) => void;
-  onCancel: () => void;
-  TooltipIcon: typeof TooltipIcon;
-  roomId: string | undefined;
-}
-
-//PageEditor component
-const PageEditor: React.FC<PageEditorProps> = ({ page, onSave, onCancel, TooltipIcon, roomId }) => {
-  const [content, setContent] = useState(page.content);
-
-  const { ttsOn } = useContext(TTSToggleContext);
-  // const { id: roomId } = useParams<{ id: string }>();
-
-  const handleContentChange = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setContent(event.target.value);
-    if(socket) {
-      socket.emit('typing', {roomId, content: event.target.value});
-    }
-  };
-
-  useEffect(() => {
-    if (socket) {
-      socket.on('typing', content => {
-        setContent(content);
-      });
-
-      return () => {
-        //cleanup from typing
-        socket.off('typing');
-      };
-    }
-  }, [content]);
-
-  const handleSave = () => {
-    onSave(content);
-  };
-
-  const handleCancel = () => {
-    onCancel();
-  };
-
-  //for transcribing data into the page
-  const updateContentWithTranscript = (transcript: string) => {
-    setContent((prevContent) => prevContent + ' ' + transcript);
-  };
-
-  const handleSpeakClick = () => {
-    if (page && 'speechSynthesis' in window) {
-      //cancel any other ongoing speech synthesis
-      window.speechSynthesis.cancel();
-      const utterance = new SpeechSynthesisUtterance(page.content);
-      window.speechSynthesis.speak(utterance);
-    } else {
-      console.error('Speech synthesis is not supported in this browser.');
-    }
-  };
-
-  const handleClearContent = () => {
-    setContent('');
-  };
-
-
-  return (
-    <div>
-      <div style={{ position: 'relative', display: 'inline-block', top: '100px' }}>
-        <GrammarlyEditorPlugin clientId='client_RZvMQYBxstbSeZEA6Ft7sA'>
-          <textarea
-            value={ content }
-            onChange={ handleContentChange }
-            maxLength={310}
-            rows={10}
-            cols={50}
-            style={{ width: '100%', height: '50px' }}
-          />
-        </GrammarlyEditorPlugin>
-        <FaTimesCircle
-          style={{
-            position: 'absolute',
-            top: '-30px',
-            right: 0,
-            color: '#3d3983',
-            backgroundColor: 'white',
-            borderRadius: '100%'
-          }}
-          size={30}
-          onClick={ handleCancel }
-        />
-      <div style={{
-        position: 'absolute',
-        bottom: '-60px',
-        right: 0,
-        display: 'flex',
-        gap: '10px',
-      }}>
-        <TooltipIcon
-          icon={ FaSave }
-          tooltipText="Save"
-          handleClick={ handleSave }
-          style={{ top: '15px'}}
-        />
-          <TooltipIcon
-            icon={ FaEraser }
-            tooltipText="Clear"
-            handleClick={ handleClearContent }
-            style={{ top: '15px'}}
-          />
-        <TooltipIcon
-          icon={ FaVolumeUp }
-          tooltipText="TTY"
-          handleClick={ handleSpeakClick }
-          style={{ top: '15px'}}
-        />
-        <STT updateTranscript={ updateContentWithTranscript } />
-      </div>
-    </div>
-  </div>
-  );
-
-};
 
 interface FlipBookProps {
   story: Story;
@@ -261,33 +130,6 @@ const FlipBook: React.FC<FlipBookProps> = ({ story, selectedStoryPages, onUpdate
     }
   };
 
-
-  // const handleToggleAutoRead = () => {
-  //   setIsAutoReading(!isAutoReading);
-  //   if (window.speechSynthesis.speaking) {
-  //     if (isAutoReading) {
-  //       window.speechSynthesis.pause();
-  //     } else {
-  //       window.speechSynthesis.resume();
-  //     }
-  //   } else if (isAutoReading) {
-  //     handleSpeakClick(selectedStoryPages[flipBookRef.current.pageIndex]?.content);
-  //   }
-  // };
-
-  // const handleImageUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
-  //   const file = event.target.files?.[0];
-  //   if (file) {
-  //     const url = URL.createObjectURL(file);
-  //     // Here we set the coverImage of the current story. This will be a local URL.
-  //     // Remember, this URL will not be persistent and it will only be available in the current session.
-  //     // You will have to send this image to the server to store it and get a persistent URL in a real application.
-  //     // Also, remember to revoke this URL when it's not needed anymore.
-  //     story.coverImage = url;
-  //   }
-  // }
-
-
   return (
     <>
     <div style={{display: 'flex', justifyContent: 'space-evenly'}}>
@@ -315,11 +157,6 @@ const FlipBook: React.FC<FlipBookProps> = ({ story, selectedStoryPages, onUpdate
       startPage={1}
       showCover={true}
       useMouseEvents={true}
-      // onFlip={(e: any) => {
-      //   if (isAutoReading) {
-      //     handleSpeakClick(selectedStoryPages[e.data]?.content);
-      //   }
-      // }}
       style={{
         backgroundColor: '#fbf5df',
         boxShadow: "0 2px 5px rgba(0, 0, 0, 0.3)",
